pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "topology_generator"
version = "0.1.0"
description = "Generate, visualize, and export datacenter network topologies."
authors = [
    { name = "Hitesh Kumar", email = "hitesh.kumar@computacenter.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib",
    "networkx",
    "numpy",
    "openpyxl",
    "pandas",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "ruff",
    "types-PyYAML",
]

[project.scripts]
topology-generator = "topology_generator.main:main"

[tool.setuptools.packages.find]
include = ["topology_generator*"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".git-rewrite",
    ".pyenv",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
packages = ["topology_generator"]
check_untyped_defs = true
exclude = ["tests/"]
no_implicit_optional = true
pretty = true
show_error_codes = true
warn_redundant_casts = true
warn_unused_ignores = true
